Integrated AXIS-CSRM ontologies
The class `TDBManager` is set up to manage connections to a TDB store. And
the OWL files containing AXIS-CSRM v0.4 are now embedded. The `TDBManager`
even has a sample `main` function allowing anyone to test the new
features.

Please note that all file names and folder paths are absolute and
harcoded, and will be replaced by a more suitable solution as soon as
possible.