Working path /register/categories/ in servlet
Added a repository into maven :
    <dependencies>
        <!--  Gson: Java to Json conversion -->
        <dependency>
          <groupId>com.google.code.gson</groupId>
          <artifactId>gson</artifactId>
          <version>2.8.0</version>
          <scope>compile</scope>
        </dependency>
        <!-- [...] -->
    </dependencies>

This is used to have a good user friendly json parser/serializer.

For more details about this json, please have a look at https://github.com/google/gson

Also, for the moment, the servlet uses the URI to know which method to
use, this might be changed later.