Deployment fixes
GlassFish is a shitty deployment server which doesn't use the same
working directory on Windows, Linux or macOS. So, the setup specific
structures (such as the upload folder, or the TDB folder) is now
automatically done when required, as it always should have been.

Moreover, please note that for the server to be able to find the
ontologies, the `resources` folder needs to be copied in GlassFish's
working directory. Otherwise, you will be faced with some random
`NullPointerException` coming out of the blue.