Solved performance issues
As per issue #2, heavy performance issues were reported whenever a
inferred model was used.

After investigations, it seems that the default reasoner makes **a lot**
of useless entailments, filling the memory and making the model
unusable.

The default reasoner has now been replaced with the *micro* reasoner. It
meets AXIS-SOW-POC's needs for the moment, and makes the reasoning
**way** faster. For further information, take a look at
[Jena's documentation about reasoners](https://jena.apache.org/documentation/inference/#owl-coverage).